% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mvrnorm_ma.R
\name{mvrnorm_ma}
\alias{mvrnorm_ma}
\title{Simulate from a Multivariate Normal Distribution with MA(q) dependence}
\usage{
mvrnorm_ma(n = 1, mu, Sigma, rho = 0, tol = 1e-06, empirical = FALSE)
}
\arguments{
\item{n}{the number of samples required.}

\item{mu}{a vector giving the means of the variables.}

\item{Sigma}{a positive-definite symmetric matrix specifying the
    covariance matrix of the variables.}

\item{rho}{A numeric vector of MA coeficients.}

\item{tol}{tolerance (relative to largest variance) for numerical lack
    of positive-definiteness in \code{Sigma}.}

\item{empirical}{logical. If true, mu and Sigma specify the empirical
    not population mean and covariance matrix.}
}
\value{
A matrix
}
\description{
Produces one or more samples from the specified multivariate normal distribution,
in which the errors are MA(q).
}

