% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pols503-package.R
\name{r2_to_sigma}
\alias{r2_to_sigma}
\title{Find regression standard error to produce a given R-squared}
\usage{
r2_to_sigma(X, b, r2)
}
\arguments{
\item{X}{The sample design matrix}

\item{b}{The population coefficients for X}

\item{r2}{The desired R^2}
}
\value{
The regression standard deviation to produce that R^2.
}
\description{
For a normal linear model, given a sample X and population beta,
find the regression standard deviation to produce the desired R-squared.
}

